+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 > - **Security** 对安全的改进
 
 ## [v5.1.3.100] 
+### Removed
+  - 删除冗余表ntc_message_to_user,ntc_messages,ntc_weather 
 ### Security
   - 统一处理接口参数无效异常，解决安全扫描问题：JSON 中反映的未清理用户输入